fix: CODAP case creation notification listener not setup when plugin starts as collapsed [PT-185324855]
This fixes an issue with the case creation notification not running when this plugin is included in a CODAP document and the plugin is collapsed (minimized) at load time.  This manifested as predictions not being updated when new cases were created in the testing table in the "Exploring AI in ELA with StoryQ (SY 2022)" sequence as the plugin was collapsed on page 4 in the first activity.

The tab component in the root Storyq component was not rendering the TestingPanel when the plugin was collapsed and the TestingPanel setup the CODAP case creation listener.

This fix pulls the testing manager instance creation and case creation listener into the root component and passes them down to the TestingPanel.  This allows the case creation notification to be handled when the plugin is collapsed.