feat: Extract phrases and emoticons [PT-188319600] [PT-188319636]
This updates the plugin to pass the full trimmed text when pattern matching text and then changes the word boundary metacharacters in the pattern to be optional if the text begins/ends with non-word characters.

This also adds a missing escape of any regex characters in the free form text and changes spaces to match multiple spaces and expands quotes to "smart" quotes.

Finally this also disables the word highlighting as this doesn't work for phrases or existing word list features.