fix: Emoji key combinations don't enable the feature done button [PT-188413736]
This simplifies the text box component to remove the unneeded keyup handler that was not being called when the Mac emoji tool (Control+Command+Space) was used to enter a character.

The keyup handler is no longer needed as the valueChangeEvent is always "keyup", whereas in the original UI it used keyup and enter.